Update the timing of notifications and always lock
The Tichu game today is generally seen as sluggish. It waits 1 second after every pass and also after any other notification. This was dues to overly using setSynchronous(..., 1000).

Some notifications are just updating the game state in the background, and there is nothing (or not much) that the user can see and parse. For example "hasBomb" or "tichuBet" (when just signaling that a Tichu is not possible anymore). So waiting a couple of seconds even when everyone auto-passed was normal.

setSynchronous() was tweaked for each notification individually.

Also, when users take action we are now always setting `lock=true`. The main benefit here is that users get immediate feedback of their action and are prevented from accidental button clicks. Especially accidental "Tichu" clicks were reported and are obviously super annoying.