Table layout refactoring
Stock of cards sets its overlap responsive to the size of the container element and to the number of cards displayed.

The "last card played" area is converted to a grid layout. That makes the list *and* the square table layout behave better in various widths.

Game panel area max-width 1500px with an auto margin for centering.