New Layout
## Redesign the player info in the "last played" blocks

- Larger fonts
- Text overflow of player names
- Strict 150px width
- Slightly smaller Tichu bet icons
- Added point count

## Table layout refactoring

Stock of cards sets its overlap responsive to the size of the container element and to the number of cards displayed.

The "last card played" area is converted to a grid layout. That makes the list *and* the square table layout behave better in various widths.

Game panel area max-width 1500px with an auto margin for centering.

## Fixed buttons above My Hand