Remove `player_has_bomb` from player table
And remove all associated code. It was currently unused, and we are not
planning to bring it back. Besides, the frontend could figure it out on
its own, if it really wants to know. No need to add complexity by
adding database fields, keeping them updated and sending notifications
for them.